v2

+++ models/teachers/dlinear.py
@@ -8,4 +8,4 @@ class DLinear(nn.Module):
         self.fc = nn.Linear(lookback, len(horizons))
 
     def forward(self, x):
-        return self.fc(x.float())
+        return self.fc(x)
